refactor: validator를 추상클래스와 추상 매소드를 사용하여 함수를 분리
